+++ NEWS.md
@@ -1,4 +1,8 @@
 # 2.1.1.9017
+* Enabled argument `target.sig.index` to also take the signature id as input in
+function `SignaturePresenceTest1`. Changed argument `m.opts` to use return value
+from `DefaultManyOpts` by default.
+
 * Updated exported function `RareSignatures` to include SBS91 and SBS94
 as rare signatures.